Bitset translated to work on the order of bytes.

+++ include/uint_array/hybrid.hpp
@@ -178,7 +178,7 @@ namespace hybrid {
     }
     return result;
   }
-  inline size_t intersect_a16_bs(uint8_t *C_in, const unsigned short *A, const unsigned short *B, const size_t s_a, const size_t s_b) {
+  inline size_t intersect_a16_bs(uint8_t *C_in, const unsigned short *A, const uint8_t *B, const size_t s_a, const size_t s_b) {
     #if WRITE_VECTOR == 0
     (void) C_in;   
     #endif 
@@ -210,7 +210,7 @@ namespace hybrid {
     return count;
   }
   //untested
-  inline size_t intersect_a32_bs(uint8_t *C_in, const uint32_t *A, const unsigned short *B, const size_t s_a, const size_t s_b) {
+  inline size_t intersect_a32_bs(uint8_t *C_in, const uint32_t *A, const uint8_t *B, const size_t s_a, const size_t s_b) {
     #if WRITE_VECTOR == 0
     C_in[0] = common::ARRAY32;
     (void) C_in;
